+++ packages/server/db/migrations.js
@@ -1,27 +1,51 @@
 const { initializeState } = require("@bwi/shared/utils");
+const { DateTime } = require("luxon");
+const fs = require("fs");
 const { ConfigDB } = require("./level");
 const { stateTypes, amountPerFaculty } = require('@bwi/shared/constants');
 
+const stateDefinedByTask = (stateType) => {
+  return {
+    bookingMaxPerEntry: amountPerFaculty[stateType],
+    minuteInterval: 10,
+    dayEndsAt: "17:50",
+    availableDayFrom:
+      DateTime.fromObject({ year: 2021, month: 8, day: 16 }).toString(),
+    availableDayTo:
+      DateTime.fromObject({ year: 2021, month: 9, day: 16 }).toString(),
+    exclusiveDates: ["23-08-2021", "24-08-2021"],
+    inclusiveDates: ["28-08-2021"],
+    filterRules: ["weekend:2"],
+  }
+};
+
 const levelMigrate = async () => {
-  const ISWBMPEVal = (faculty) => initializeState({ bookingMaxPerEntry: amountPerFaculty[faculty], minuteInterval: 15 });
-  const { PHILOLOGICAL_FACULTY, LAW_FACULTY, FIAT, HISTORY_FACULTY, FHBBT, ECONOMICS_FACULTY, FIM } = stateTypes;
-  await ConfigDB.setConfig(PHILOLOGICAL_FACULTY, ISWBMPEVal(PHILOLOGICAL_FACULTY));
-  await ConfigDB.setConfig(FIM, ISWBMPEVal(FIM));
-  await ConfigDB.setConfig(LAW_FACULTY, ISWBMPEVal(LAW_FACULTY));
-  await ConfigDB.setConfig(FIAT, ISWBMPEVal(FIAT));
-  await ConfigDB.setConfig(HISTORY_FACULTY, ISWBMPEVal(HISTORY_FACULTY));
-  await ConfigDB.setConfig(FHBBT, ISWBMPEVal(FHBBT));
-  await ConfigDB.setConfig(ECONOMICS_FACULTY, ISWBMPEVal(ECONOMICS_FACULTY));
+  await ConfigDB.clearConfig();
+  const ISWBMPEVal = (stateType) => initializeState(stateDefinedByTask(stateType));
+  const { LAWYERS } = stateTypes;
+  await ConfigDB.setConfig(LAWYERS, ISWBMPEVal(LAWYERS));
+  console.log("LEVEL DB SUCCESSFULLY INITIALIZED.");
 };
 
 const squirrelMigrate = async () => {
-  console.log('TODO');
+  const demons = __dirname + '/db.sqlite3';
+  let err;
+  if (fs.existsSync(demons)) {
+    err = await new Promise(r =>
+      fs.unlink(__dirname + '/db.sqlite3', (err) => r(err)));
+  }
+  if (err && err.errno !== -2) {
+    throw new Error("Something happened when removing db");
+  }
+  fs.writeFileSync(__dirname + '/db.sqlite3', "");
+  const { initializeDB } = require("./db");
+  initializeDB(true, false).then(() => {
+    console.log("SQLITE DB SUCCESSFULLY INITIALIZED.");
+    process.exit(0);
+  });
 };
 
-// levelMigrate().then(() => {
-//   console.log('flex');
-//   ConfigDB.getConfig(stateTypes.PHILOLOGICAL_FACULTY).then((a) => {
-//     console.log(a);
-//   });
-// });
-levelMigrate();
+(async () => {
+  await levelMigrate();
+  await squirrelMigrate();
+})();