+++ packages/server/server.js
@@ -1,36 +1,24 @@
+require('dotenv').config();
 const express = require('express');
 const path = require('path');
 const cors = require('cors');
-// const history = require('connect-history-api-fallback');
-const { v4: uuid } = require('uuid');
 const morgan = require('morgan');
-// const expressSession = require("express-session");
+// const { v4: uuid } = require('uuid');
+const expressSession = require("express-session");
+const store = new expressSession.MemoryStore();
 // const passport = require("passport");
 
-const sessions = [];
 const app = express();
-const port = parseFloat(process.env.PORT) || 3000;
+const port = process.env.PORT || 3000;
 
 /*
 https://auth0.com/blog/create-a-simple-and-secure-node-express-app/
-
-const session = {
-  secret: process.env.SESSION_SECRET,
-  cookie: {},
-  resave: false,
-  saveUninitialized: false
-};
-
-if (app.get("env") === "production") {
-  // Serve secure cookies, requires HTTPS
-  session.cookie.secure = true;
-}
-
 */
 
 const indexRoute = require('./routers/index');
 const rootRoute = require('./routers/root');
 const stateRoute = require('./routers/states');
+const authRoute = require('./routers/auth');
 
 const whitelist = ["http://localhost:3000", "http://localhost:8080"];
 let development = true;
@@ -49,9 +37,15 @@ if (development) {
     },
   }));
 }
-// app.use(history({
-//   index: "/index.html"
-// }));
+
+app.use(expressSession({
+  secret: process.env.SESSION_SECRET || "SECRET",
+  cookie: {
+    maxAge: 30*60*100
+  },
+  saveUninitialized: false,
+  store,
+}));
 app.use(express.json({ extended: true }));
 app.use(express.urlencoded());
 app.set('views', path.join(__dirname, 'views'));
@@ -61,31 +55,15 @@ app.use(morgan(':method :url :status :res[content-length] - :response-time ms'))
 
 app.use('/api', indexRoute);
 app.use('/api', stateRoute);
-app.use('/', rootRoute);
-app.get('/check', (_, res) => {
-  res.send('Ohaio');
-});
-app.post('/postcheck', (req, res) => {
-  res.send(req.body);
-});
-
-app.post('/login', (req, res) => {
-  const { username, password } = req.body;
-  if (!username || !password) return res.sendStatus(401);
-  let uid = uuid();
-  if (username === 'jeremy' && password === 'dunk898') {
-    sessions.push(uid);
-  }
-  res.send(uid);
-});
-app.post('/valpas', (req, res) => {
-  if (req.body.pass === 'flex') {
-    return res.sendStatus(200);
-  }
-  return res.sendStatus(401);
-});
+app.use('/api', rootRoute);
+app.use('/api/auth', authRoute);
 
 // localhost:3000 -> serverIP:3000
-app.listen(port, '0.0.0.0', () => {
-  console.log('Listening on port %s', port);
-});
+let fireLock = false;
+
+if (!fireLock)
+  app.listen(port, '0.0.0.0', () => {
+    console.log('Listening on port %s', port);
+  });
+
+module.exports.app = app;