+++ packages/webapp/src/App.vue
@@ -1,9 +1,15 @@
 <template>
+  <Modal />
   <router-view></router-view>
 </template>
 
 <script>
+import Modal from "@/components/Modal";
+
 export default {
-  name: "App"
+  name: "App",
+  components: {
+    Modal,
+  },
 }
 </script>