+++ packages/server/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/ruxxzebre/booking_widget/compare/v1.4.3...v1.5.0) (2021-08-10)
+
+
+### Features
+
+* simple authentication ([cf5b9de](https://github.com/ruxxzebre/booking_widget/commit/cf5b9de40ea0599da652734012a8c01ebb444c3a))
+
+
+
+
+
+## [1.4.3](https://github.com/ruxxzebre/booking_widget/compare/v1.4.1...v1.4.3) (2021-08-10)
+
+
+### Bug Fixes
+
+* add kvdb to gitignore ([0bd79f0](https://github.com/ruxxzebre/booking_widget/commit/0bd79f03c85ed4070065b04fb6b9997b15f20c42))
+* exclusive dates comparison + dates itself ([fb4eb9c](https://github.com/ruxxzebre/booking_widget/commit/fb4eb9ce0a5883677750c79db53175a2b2786af1))
+* remove kvdb.sqlite3 ([ad72e44](https://github.com/ruxxzebre/booking_widget/commit/ad72e44635d42a79a9892bf73b816c2b56b752f2))
+* versions ([3bd0ba5](https://github.com/ruxxzebre/booking_widget/commit/3bd0ba5da5bdb37e98124bbd97b312b73a565496))
+
+
+
+
+
+## [1.4.1](https://github.com/ruxxzebre/booking_widget/compare/v1.4.0...v1.4.1) (2021-08-10)
+
+**Note:** Version bump only for package @bwi/server
+
+
+
+
+
+# [1.4.0](https://github.com/ruxxzebre/booking_widget/compare/v1.3.0...v1.4.0) (2021-08-10)
+
+
+### Bug Fixes
+
+* cleanup ([2513276](https://github.com/ruxxzebre/booking_widget/commit/2513276522460e04e2cb75e416bbe9ce62fa6835))
+* vueconf ([8009427](https://github.com/ruxxzebre/booking_widget/commit/800942719bbadba9b83043b8d3f9299d42eea033))
+
+
+### Features
+
+* export server app ([a2d6434](https://github.com/ruxxzebre/booking_widget/commit/a2d643487739fc5a9464c08d68be07320176d15d))
+* inclusive and exclusive dates, filter rules + rules parser ([a26cecb](https://github.com/ruxxzebre/booking_widget/commit/a26cecb682e443b07ef8608d26f4c438ce8437c4))
+
+
+
+
+
+# [1.3.0](https://github.com/ruxxzebre/booking_widget/compare/v1.2.0...v1.3.0) (2021-08-09)
+
+
+### Bug Fixes
+
+* await stuff ([10f803f](https://github.com/ruxxzebre/booking_widget/commit/10f803f3f7db685ec3b83de110d39246d3449441))
+* sanitize user entry data ([7d21613](https://github.com/ruxxzebre/booking_widget/commit/7d216133011dd4d1d60e23cc6ba6b4f214fb0d6c))
+
+
+### Features
+
+* complete migrations ([68cd16c](https://github.com/ruxxzebre/booking_widget/commit/68cd16cc6cbba95e989b92532aae9c0608f00805))
+* Docker Compose + Dockerfiles ([9700802](https://github.com/ruxxzebre/booking_widget/commit/9700802b5b51f00763102c9c3b8853494ee9b3d6))
+* entry existance enchanced checker + error/success messages handling. ([c526c3e](https://github.com/ruxxzebre/booking_widget/commit/c526c3e4b26b488ed981b0101b369111596e6e23))
+* keyv + clearConfig ([92e4770](https://github.com/ruxxzebre/booking_widget/commit/92e4770ce29a455f2765fa8f44c1fdf8fd5a526d))
+* migrations according to task ([62e8126](https://github.com/ruxxzebre/booking_widget/commit/62e81268418bf5a34446ec5a499225b6dc6ae33a))
+
+
+
+
+
+# [1.2.0](https://github.com/ruxxzebre/booking_widget/compare/v1.1.0...v1.2.0) (2021-08-05)
+
+
+### Features
+
+* lawyers flex ([44d3b98](https://github.com/ruxxzebre/booking_widget/commit/44d3b984cc0e09070bbacceafcd096bc1de4d6c9))
+
+
+
+
+
 # [1.1.0](https://github.com/ruxxzebre/booking_widget/compare/v1.0.1...v1.1.0) (2021-08-05)