+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/ruxxzebre/booking_widget/compare/v1.4.0...v1.4.1) (2021-08-10)
+
+**Note:** Version bump only for package @bwi/shared
+
+
+
+
+
+# [1.4.0](https://github.com/ruxxzebre/booking_widget/compare/v1.3.0...v1.4.0) (2021-08-10)
+
+
+### Bug Fixes
+
+* vueconf ([8009427](https://github.com/ruxxzebre/booking_widget/commit/800942719bbadba9b83043b8d3f9299d42eea033))
+
+
+### Features
+
+* inclusive and exclusive dates, filter rules + rules parser ([a26cecb](https://github.com/ruxxzebre/booking_widget/commit/a26cecb682e443b07ef8608d26f4c438ce8437c4))
+
+
+
+
+
+# [1.3.0](https://github.com/ruxxzebre/booking_widget/compare/v1.2.0...v1.3.0) (2021-08-09)
+
+
+### Bug Fixes
+
+* unavaliable times right ([c39f503](https://github.com/ruxxzebre/booking_widget/commit/c39f503a0ba209e4f9396593c0233bc4a76c9952))
+
+
+### Features
+
+* Docker Compose + Dockerfiles ([9700802](https://github.com/ruxxzebre/booking_widget/commit/9700802b5b51f00763102c9c3b8853494ee9b3d6))
+* migrations according to task ([62e8126](https://github.com/ruxxzebre/booking_widget/commit/62e81268418bf5a34446ec5a499225b6dc6ae33a))
+
+
+
+
+
+# [1.2.0](https://github.com/ruxxzebre/booking_widget/compare/v1.1.0...v1.2.0) (2021-08-05)
+
+
+### Features
+
+* lawyers flex ([44d3b98](https://github.com/ruxxzebre/booking_widget/commit/44d3b984cc0e09070bbacceafcd096bc1de4d6c9))
+
+
+
+
+
 # [1.1.0](https://github.com/ruxxzebre/booking_widget/compare/v1.0.1...v1.1.0) (2021-08-05)