+++ packages/webapp/CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/ruxxzebre/booking_widget/compare/v1.4.3...v1.5.0) (2021-08-10)
+
+
+### Bug Fixes
+
+* admin page needs store for smth, patch it up for a moment ([b644357](https://github.com/ruxxzebre/booking_widget/commit/b64435750ec4d38965d61c2da2a6640eac0b8c52))
+* modal error mode layout ([b234174](https://github.com/ruxxzebre/booking_widget/commit/b23417479f6ae5e61c7112e6aa976a0cd6fec7d0))
+* simplest admin page ([33891af](https://github.com/ruxxzebre/booking_widget/commit/33891afc43c1df010f045e6949fd59d8d2de01bf))
+
+
+### Features
+
+* new modal ([7c75562](https://github.com/ruxxzebre/booking_widget/commit/7c75562cbea021064d798c93d072d042a28963e5))
+* todo separated modal component ([301baf7](https://github.com/ruxxzebre/booking_widget/commit/301baf7541a5a20c1519300341ce6ff3a324b67c))
+
+
+
+
+
+## [1.4.3](https://github.com/ruxxzebre/booking_widget/compare/v1.4.1...v1.4.3) (2021-08-10)
+
+
+### Bug Fixes
+
+* exclusive dates comparison + dates itself ([fb4eb9c](https://github.com/ruxxzebre/booking_widget/commit/fb4eb9ce0a5883677750c79db53175a2b2786af1))
+* versions ([3bd0ba5](https://github.com/ruxxzebre/booking_widget/commit/3bd0ba5da5bdb37e98124bbd97b312b73a565496))
+
+
+
+
+
+## [1.4.1](https://github.com/ruxxzebre/booking_widget/compare/v1.4.0...v1.4.1) (2021-08-10)
+
+**Note:** Version bump only for package @bwi/webapp
+
+
+
+
+
+# [1.4.0](https://github.com/ruxxzebre/booking_widget/compare/v1.3.0...v1.4.0) (2021-08-10)
+
+
+### Bug Fixes
+
+* spa ([aacdc0d](https://github.com/ruxxzebre/booking_widget/commit/aacdc0d43b22acdefbc90473fa87f9c74756fd8d))
+* sugarcrash ([b523932](https://github.com/ruxxzebre/booking_widget/commit/b5239325e7c38374051b2d25ecad492050933f5b))
+* vueconf ([8009427](https://github.com/ruxxzebre/booking_widget/commit/800942719bbadba9b83043b8d3f9299d42eea033))
+
+
+### Features
+
+* easter 1 ([439296c](https://github.com/ruxxzebre/booking_widget/commit/439296cba375a291ee395efeb7d10644901b44c7))
+* inclusive and exclusive dates, filter rules + rules parser ([a26cecb](https://github.com/ruxxzebre/booking_widget/commit/a26cecb682e443b07ef8608d26f4c438ce8437c4))
+
+
+
+
+
+# [1.3.0](https://github.com/ruxxzebre/booking_widget/compare/v1.2.0...v1.3.0) (2021-08-09)
+
+
+### Bug Fixes
+
+* cleanup ([cb1419b](https://github.com/ruxxzebre/booking_widget/commit/cb1419b1c7c60dd3234d49f5c08e6448522a01b3))
+* fetching entries ([a38a98c](https://github.com/ruxxzebre/booking_widget/commit/a38a98cbf5790a9288b1f570450351c649723cfd))
+* generating entries ([5bf3f56](https://github.com/ruxxzebre/booking_widget/commit/5bf3f566e1fd62b2d9d295e0f0b522fdf4b0b4b3))
+* generating entries + cleanup ([75337dd](https://github.com/ruxxzebre/booking_widget/commit/75337dd8c3e5df05af09b9f8cc37f0cd9b3a16da))
+* patch persistance ([b9a0095](https://github.com/ruxxzebre/booking_widget/commit/b9a0095a33a365e93b187dd956c247624fb705a3))
+* remove console.log ([38cb9b7](https://github.com/ruxxzebre/booking_widget/commit/38cb9b757eeceb744673920eb3428cce903f05f9))
+* remove debug times ([4ad6cf5](https://github.com/ruxxzebre/booking_widget/commit/4ad6cf541c39132b20f7f0adb8e50763294ff865))
+* remove unnecessary mutations + fix adding entry + fix fetching entries ([bf95ec7](https://github.com/ruxxzebre/booking_widget/commit/bf95ec736567b3e0179e4b1212ff8cfde198adb7))
+* state in addEntry ([8b1f6b8](https://github.com/ruxxzebre/booking_widget/commit/8b1f6b8953f74e6f1f5acf40a5818780f2b2e254))
+
+
+### Features
+
+* Docker Compose + Dockerfiles ([9700802](https://github.com/ruxxzebre/booking_widget/commit/9700802b5b51f00763102c9c3b8853494ee9b3d6))
+* entry existance enchanced checker + error/success messages handling. ([c526c3e](https://github.com/ruxxzebre/booking_widget/commit/c526c3e4b26b488ed981b0101b369111596e6e23))
+* fetch entries every interval ([2566879](https://github.com/ruxxzebre/booking_widget/commit/256687942a4496a59b529b21d192541ae8d0f7b3))
+
+
+
+
+
+# [1.2.0](https://github.com/ruxxzebre/booking_widget/compare/v1.1.0...v1.2.0) (2021-08-05)
+
+
+### Features
+
+* lawyers flex ([44d3b98](https://github.com/ruxxzebre/booking_widget/commit/44d3b984cc0e09070bbacceafcd096bc1de4d6c9))
+
+
+
+
+
 # [1.1.0](https://github.com/ruxxzebre/booking_widget/compare/v1.0.1...v1.1.0) (2021-08-05)